+++ core/src/lib.rs
@@ -1,6 +1,6 @@
 #![crate_type = "proc-macro"]
 #![recursion_limit = "256"]
-#![doc(html_root_url = "https://docs.rs/auto_enums_core/0.1.3")]
+#![doc(html_root_url = "https://docs.rs/auto_enums_core/0.2.0")]
 
 extern crate proc_macro;
 extern crate proc_macro2;